+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.2.5 - 2024-08-06
 ### Changed
 - Confirm box to prevent deleting all files of a duplicate
+- Userid is now automatically set when inserting or updating an entity
 
 ## 1.2.4 - 2024-07-24
 ### Added